requirements.txt
django-axes  
django-admin-interface
mercadopago
django-simple-captcha